Fix Java implementations of cast_to_uint/1.
library/uint8.m:
library/uint16.m:
library/uint32.m:
library/uint64.m:
     Deal with the XXXs in the Java foreign_procs for cast_to_uint/1

+++ library/uint64.m
@@ -457,8 +457,7 @@ cast_to_int(_) = _ :-
     cast_to_uint(U64::in) = (U::out),
     [will_not_call_mercury, promise_pure, thread_safe],
 "
-    // XXX This is probably wrong.
-    U = U64;
+    U = (int) U64;
 ").
 
 :- pragma no_determinism_warning(cast_to_uint/1).
